pyproject.toml
[tool.poetry]
name = "kitchen-robot"
version = "0.1.0"
description = ""
authors = ["Isma-Ilou Sadou <ismailou.sa@gmail.com>"]
readme = "README.md"
packages = [{include = "*", from="src"}]

[tool.poetry.dependencies]
python = "^3.9"
python-box = "^7.1.1"
pyyaml = "^6.0.1"
ensure = "^1.0.4"
pandas = "^2.2.0"
gdown = "^5.1.0"
torch = "^2.2.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
ruff = "^0.2.2"
pytest = "^8.0.2"
notebook = "^7.1.0"
scikit-learn = "^1.4.1.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
